Refactoring core project
