+++ MVCProxy.Core/Classes/ProxyHttpRoute.cs
@@ -1,15 +1,9 @@
 ﻿using System;
 
-namespace MVCProxy.Core.View
+namespace MVCProxy.Core.Classes
 {
     public class ProxyHttpRoute
     {
-        public ProxyHttpRoute()
-        {
-            Defaults = new { };
-            Constraints = new { };
-        }
-
         public string Name { get; set; }
         public string RouteTemplate { get; set; }
         public Object Defaults { get; set; }
