+++ MVCProxy.Core/Extenssions/HttpConfigurationExtensions.cs
@@ -1,14 +1,14 @@
-﻿using MVCProxy.Core.View;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Net.Http;
 using System.Web.Http;
+using MVCProxy.Core.Classes;
 
-namespace MVCProxy.Core.Application
+namespace MVCProxy.Core.Extenssions
 {
     public static class HttpConfigurationExtensions
     {
         public static void EnableProxy(this HttpConfiguration httpConfig, string host, int port,
-            List<ProxyHttpRoute> proxyRoutes)
+            IEnumerable<ProxyHttpRoute> proxyRoutes)
         {
             foreach (var route in proxyRoutes)
             {
