+++ MVCProxy.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 15.0.27004.2006
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MVCProxy.Core", "MVCProxy.Core\MVCProxy.Core.csproj", "{5E227DB8-9194-4C22-83D3-955A69D9F470}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MVCProxy.Tests.API", "MVCProxy.Tests.API\MVCProxy.Tests.API.csproj", "{67E01AF4-C664-4704-922A-3BD05AB7F6DF}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -17,10 +15,6 @@ Global
 		{5E227DB8-9194-4C22-83D3-955A69D9F470}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{5E227DB8-9194-4C22-83D3-955A69D9F470}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{5E227DB8-9194-4C22-83D3-955A69D9F470}.Release|Any CPU.Build.0 = Release|Any CPU
-		{67E01AF4-C664-4704-922A-3BD05AB7F6DF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{67E01AF4-C664-4704-922A-3BD05AB7F6DF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{67E01AF4-C664-4704-922A-3BD05AB7F6DF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{67E01AF4-C664-4704-922A-3BD05AB7F6DF}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
